# Docs/archive/openAPI.yaml
openapi: 3.0.3
info:
  title: Restaurant Ordering System API
  version: 1.0.0
  description: API for managing restaurant orders, menu, tables, and payments
servers:
  - url: http://localhost:5000/api
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string }
        role:
          type: string
          enum: [admin, waiter, kitchen]
...
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }

    # Category
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
    CategoryRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }

    # MenuItem
...
security:
  - BearerAuth: []
paths:
...
  /users/{id}:
    put:
      tags: [Users]
      summary: User aktualisieren (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User nicht gefunden

    delete:
      tags: [Users]
      summary: User löschen (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: User gelöscht
        '404':
          description: User nicht gefunden

  # ─── CATEGORIES ─────────────────────────────────────────
  /categories:
    get:
      tags: [Menu]
      summary: Alle Kategorien abrufen
      security: []
      responses:
        '200':
          description: Liste der Kategorien
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      tags: [Menu]
      summary: Kategorie erstellen (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Kategorie erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /categories/{id}:
    put:
      tags: [Menu]
      summary: Kategorie aktualisieren (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Kategorie aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Kategorie nicht gefunden

    delete:
      tags: [Menu]
      summary: Kategorie löschen (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Kategorie gelöscht
        '404':
          description: Kategorie nicht gefunden

  # ─── MENU ITEMS ─────────────────────────────────────────
...